server/requirements.txt
flask
flask-sqlalchemy
boto3==1.34.109
flask-smorest
python-dotenv
cryptography
python-dotenv
pymysql
marshmallow
marshmallow_enum